D365FO_SharePoint__DefaultUploadingUser
Are you using SharePoint in Dynamics 365 Finance & Operations for document management?
What if the user can log in on SharePoint and can view all the documents?

In the small project, I introduce the default SharePoint document user concept. You need only grant the rights to one user on your SharePoint document folder, and that user will upload the document on the behalf of all other users.

So, by implementing this, no user can connect to your SharePoint and view the documents of others.

Create a separate account on your desired instance.
Log in to the environment only once, and enjoy the solution.